requirements.txt
streamlit
groq
python-dotenv
# SQLAlchemy
pypdf2
python-docx
pytesseract
Pillow
pdf2image
fastapi
uvicorn
python-multipart
# Add other database-specific drivers if aevan is not SQLite, e.g., psycopg2-binary for PostgreSQL
